Try to check if the FILE*/fd points to a file
Try to check if the FILE*/fd points to a file and if not deal with
it like stdin/stdout. Leave the lastOp as uncomitted instead of
trying to guess read/write.

Calling >>#primRead:... through a modified MultiByteFileStream does
not seem to work but calling it directly seems to do the right thing.